fix: move datasets UI component to DS UI package
