+++ workbench-core/datasets-ui/package.json
@@ -54,6 +54,7 @@
     "@types/heft-jest": "1.0.2",
     "@types/node": "^14",
     "@types/react": "17.0.38",
+    "@types/react-dom": "^18.0.6",
     "@types/uuid": "^8.3.4",
     "depcheck": "^1.4.3",
     "eslint": "^8.7.0",
