+++ workbench-core/datasets-ui/src/index.tsx
@@ -8,6 +8,7 @@ import { columnDefinitions, searchableColumns } from './datasets-table-config/da
 import { filteringOptions } from './datasets-table-config/datasetsFilteringOptions';
 import { filteringProperties } from './datasets-table-config/datasetsFilteringProperties';
 import { type DatasetItem } from './models/Dataset';
+import { DatasetsPage } from './pages/index';
 
 
 export {
@@ -19,4 +20,7 @@ export {
 
   // From '/models' folder
   DatasetItem,
+
+  // From 'pages' folder
+  DatasetsPage
 };
